[Rename] 각 부분 폴더 생성 #34
