+++ src/app/login/page.tsx
@@ -1,4 +1,10 @@
+import LoginForm from '@/features/auth/container/login-form/LoginForm';
+
 function Login() {
-  return <div>Login</div>;
+  return (
+    <div className="flex min-h-screen w-full items-center justify-center bg-gray-100">
+      <LoginForm />
+    </div>
+  );
 }
 export default Login;
