--- src/features/auth/container/index.ts
@@ -1,1 +1,0 @@
-export {};
