+++ src/features/auth/container/login-form/LoginForm.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import FormField from '../components/FormField';
-import SubmitButton from '../components/SubmitButton';
+import FormField from '../../components/form-field/FormField';
+import SubmitButton from '../../components/submit-button/SubmitButton';
 
 function LoginForm() {
   return (
     <div className="flex h-auto w-full flex-col items-center justify-center rounded-3xl bg-white px-4 py-8 sm:h-[422px] sm:w-[510px] sm:px-14">
       <div className="flex w-full flex-col items-center justify-center gap-6">
         <h3 className="text-2xl font-semibold">로그인</h3>
-        <form className="flex w-full flex-col gap-6">
+        <form className="flex w-full flex-col gap-6" data-testid="login-form">
           <FormField
             label="아이디"
             type="text"
